owner response added

+++ src/main/java/com/masai/controller/OwnerController.java
@@ -3,6 +3,8 @@ package com.masai.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,6 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.masai.entity.Owner;
 import com.masai.entity.UserType;
+import com.masai.exception.NullValueException;
 import com.masai.service.OwnerService;
 @RestController
 @RequestMapping("/Owner")
@@ -21,21 +24,26 @@ public class OwnerController {
 	private OwnerService ownerService;
 	
 	 @PostMapping("/")
-	    public Owner saveOwnerHandler(@RequestBody Owner owner) throws Exception {
-//	        return ownerService.addOwner(custDTO,UserType.Owner);
-		 return ownerService.addOwner(owner, UserType.Owner);
+	    public ResponseEntity<Owner> saveOwnerHandler(@RequestBody Owner owner) throws Exception {
+		 return new ResponseEntity<Owner>(ownerService.addOwner(owner, UserType.Owner),HttpStatus.ACCEPTED);
 	    }
 
 
 	    @DeleteMapping("/{id}")
-	    public String removeOwnerHandler(@PathVariable Integer id) {
+	    public ResponseEntity<String> removeOwnerHandler(@PathVariable Integer id) {
+	    	if(id==null) {
+	    		throw new NullPointerException("Id is invalid");
+	    	}
 	        String owner = ownerService.removeOwner(id);
-	        return "Deleted " + owner;
+	        return new ResponseEntity<String>("Deleted " + owner,HttpStatus.OK);
 	    }
 
 	    
 	    @GetMapping("/allOwner")
-		public List<Owner> getAllOwner(){
-			return ownerService.getAllOwners();
+		public ResponseEntity<List<Owner>> getAllOwner(){
+	    	if(ownerService.getAllOwners()==null) {
+				throw  new NullValueException("Owner not found");
+			}
+			return new ResponseEntity<List<Owner>>(ownerService.getAllOwners(),HttpStatus.ACCEPTED);
 		}
 }
